+++ src/models/extend.model.ts
@@ -11,11 +11,11 @@ import { IsString, IsNotEmpty } from 'class-validator';
 export class Extend {
   @IsNotEmpty()
   @IsString()
-  @prop({ required: true, validate: /\S+/ })
+  @prop({ required: false, validate: /\S+/ })
   name: string;
 
   @IsNotEmpty()
   @IsString()
-  @prop({ required: true, validate: /\S+/ })
+  @prop({ required: false, validate: /\S+/ })
   value: string;
 }